requirements.txt
streamlit>=1.33.0
requests>=2.32.3
beautifulsoup4>=4.13.0
tqdm>=4.66.0
python-dotenv>=1.0.1

openai>=1.25.0
pinecone>=6.0.0

langchain-openai>=0.0.6
langchain-pinecone>=0.0.4
langchain-core>=0.1.57
langchain-text-splitters>=0.0.1
